[HBASE-5783] [0.89-fb] Reliably bulk load Puts/Deletes to HBase, without WAL.
Author: aaiyer

Summary:
Currently, there are 2 ways to load data *reliably* into HBase:
(i) individual Puts/Gets to HBase with WAL
(ii) MapReduce based Bulk Load mechanism

Option (i) is slow due to WALs. Option (ii) is slow due to the map
reduce merges doing everything through disk.

This is a hybrid approach of writing everything, directly to memstore,
without WAL. And, then waiting/initiating a memstore flush to provide
durability.

Test Plan:
add a new unit test
 TestBatchedUpload

16 /tmp/2012-09-18_09_07_29_hbase-89int_newHiveUpload_v5-mrFailures  Failed on MR:  TestBatchedUpload TestFromClientSide TestHRegionCloseRetry TestHRegionServerFileSystemFailure TestHsHaServerCmdLine TestMasterFailover TestMultiClusters TestNativeThriftClient TestRegionSplitter TestReplication TestSchemaModificationLocks TestThreadPoolServerFramedCmdLine TestThreadPoolServerUnframedCmdLine TestThreadedSelectorServerCmdLine TestThriftServer TestTimestamp
1 /tmp/2012-09-18_09_07_29_hbase-89int_newHiveUpload_v5-localFailures  Failed on Local:  None.

Reviewers: kranganathan, liyintang

Reviewed By: kranganathan

CC: hbase-eng@, mbautin, gqchen

Differential Revision: https://phabricator.fb.com/D570100

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1387312 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/util/Threads.java
@@ -138,7 +138,7 @@ public class Threads {
   /**
    * @param millis How long to sleep for in milliseconds.
    */
-  public static void sleep(int millis) {
+  public static void sleep(long millis) {
     try {
       Thread.sleep(millis);
     } catch (InterruptedException e) {
@@ -150,7 +150,7 @@ public class Threads {
    * Sleeps for the given amount of time. Retains the thread's interruption status. 
    * @param millis How long to sleep for in milliseconds.
    */
-  public static void sleepRetainInterrupt(int millis) {
+  public static void sleepRetainInterrupt(long millis) {
     try {
       Thread.sleep(millis);
     } catch (InterruptedException e) {
